Switch to using a namespace module for archiving
Since we will be adding extra models for archived petitions
like debate outcomes and government responses switch to using
a module for namespacing.
