+++ app/models/archived.rb
@@ -1,0 +1,7 @@
+module Archived
+  class << self
+    def table_name_prefix
+      "archived_"
+    end
+  end
+end
